Implement scroll positioning
Because the list is not a built in select tag, we must manage scroll position manually. This introduces substantial complexity.

There's a lot of compontent state here, and it's being mutated all over the place. Is there a pattern for this (the elm architecture? xstate?) that would help keep things sane?