Fix input width changing
OOooooooh boy this took forever. On typing into the input, the width of the input would change. It took a lot of fiddling, but I _think_ the problem was there was nothing up the chain setting an actual width. So now index.svelte sets it to 100%, still with a max-width 448px (this is the tailwind 'md' size).

While at it, also fixed the scrollbar styling. Turns out one must set a property on -webkit-scrollbar to set properties on -webkit-scrollbar-track or -webkit-scrollbar-thumb.