Add nearby emotions
This did not take too much work, and I consider it 'hacked together'. The actual content needs revisiting (in a sister python project) and I'm not sure the aria-controls annotation is sufficient.

This component continues to grow, and I should at least split it into combobox, entry and intro section sub-components.